safari-600.7-branch/LayoutTests/platform/ios-sim/editing/secure-input/password-input-changed-type-expected.txt

Verify that changing a password input's type updates secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


A password input is focused:
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.

After changing the type to text:
PASS testRunner.secureEventInputIsEnabled is false

After changing the type back to password:
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
